fix(signing): validate full-width permission IDs
Preserve Transaction.Contract.Permission_id as int32 throughout the
transaction review model and reject values outside the supported 0..9 range
before selecting the address display path.

Prevent values such as 256, 258, and -256 from aliasing valid permission IDs
through uint8 truncation while the original transaction bytes are signed.

Add regression coverage for invalid signed boundaries, the exact 256 protobuf
encoding, and successful review and signing with permission ID 9.

+++ src/parse.h
@@ -171,7 +171,7 @@ typedef struct txContent_t {
     uint32_t customSelector;
     contractType_e contractType;
     uint64_t dataBytes;
-    uint8_t permission_id;
+    int32_t permission_id;
     // Transaction.raw.contract is repeated on the wire, while this app can
     // safely review exactly one contract. Track the occurrence across every
     // APDU in the signing session, not only within one nanopb decode.
